Create README.md
This repository contains an exploratory data analysis (EDA) of appointment records for Brazilian public hospitals. The dataset includes various attributes of patients and indicates whether the patients showed up for their scheduled appointments. The primary goal of this analysis is to uncover trends and factors that influence patients' attendance or non-attendance at their appointments.